Rename JP full plugin install components package
